+++ lib/mumukit.rb
@@ -16,11 +16,11 @@ require_relative 'mumukit/version'
 require_relative 'mumukit/stub'
 require_relative 'mumukit/with_tempfile'
 require_relative 'mumukit/with_command_line'
-require_relative 'mumukit/file_test_compiler'
-require_relative 'mumukit/file_test_runner'
+require_relative 'mumukit/test_compiler/file_test_compiler'
+require_relative 'mumukit/test_runner/file_test_runner'
 
 require_relative 'stubs/expectations_runner'
 require_relative 'stubs/feedback_runner'
 
-require_relative 'mumukit/test_server'
-require_relative 'mumukit/test_server_app'
+require_relative 'mumukit/server/test_server'
+require_relative 'mumukit/server/test_server_app'
