+++ lib/mumukit/server/test_server.rb
