+++ lib/mumukit/server/test_server_app.rb
