+++ lib/mumukit/test_compiler/file_test_compiler.rb
